build in local dir, not github repo
Formerly, the build example used the Dockerfile in the github repo.
However, this uses the entire repo and the context folder for build.
This will not work, and we need to use the local app folder as the
context folder.

Need to figure out a good way to copy the Dockerfile and example config
file if not specified and build; maybe a bash script??